feat: udpate model structure
