+++ keychain/api/__init__.py
@@ -1,2 +1,4 @@
-from keychain.api.field import FieldModelApi as FieldModelApi
-from keychain.api.password import PasswordModelApi as PasswordModelApi
+from keychain.api.field.api import FieldModelApi
+from keychain.api.password.api import PasswordModelApi
+
+__all__ = ["PasswordModelApi", "FieldModelApi"]
