+++ keychain/api/field/__init__.py
@@ -1,0 +1,3 @@
+from keychain.api.field.api import FieldModelApi
+
+__all__ = ["FieldModelApi"]
