+++ keychain/api/password/__init__.py
@@ -1,0 +1,3 @@
+from keychain.api.password.api import PasswordModelApi
+
+__all__ = ["PasswordModelApi"]
