+++ keychain/api/password/api.py
@@ -3,28 +3,12 @@ from logging import getLogger
 from flask_appbuilder.api import ModelRestApi
 from flask_appbuilder.models.sqla.filters import FilterEqualFunction
 from flask_appbuilder.models.sqla.interface import SQLAInterface
-from marshmallow import fields, Schema
 
 from keychain.api.get_user_id import get_user_id
+from keychain.api.password.schemas import PasswordGetSchema
 from keychain.models import Password
 
 logger = getLogger(__name__)
-
-
-class FieldGetSchema(Schema):
-    id = fields.Int()
-    name = fields.Str()
-    get_value = fields.Str()
-    is_deleted = fields.Bool()
-    created_at = fields.DateTime()
-
-
-class PasswordGetSchema(Schema):
-    id = fields.Int()
-    name = fields.Str()
-    image_url = fields.Str()
-    created_at = fields.DateTime()
-    fields = fields.Nested(FieldGetSchema, many=True)
 
 
 class PasswordModelApi(ModelRestApi):
